refactor: reorganize Home component into dedicated directory
  - Move Home.tsx to Home/index.tsx for better module organization
  - Relocate related components (RecentFiles, WorkspaceManagement, WorkspaceSelector) to Home directory
